src/embeddings_app/requirements.txt
apispec==6.4.0
apispec-webframeworks==1.0.0
flasgger==0.9.7.1
Flask==3.0.1
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
GPUtil==1.4.0
llama-index==0.9.39
marshmallow==3.20.2
PyJWT==2.8.0
python-dotenv==1.0.1
retry==0.9.2
transformers==4.37.1
whatismyip
einops==0.8.0
sentence-transformers==2.7.0
